+++ resources/js/Pages/ImportStatus.vue
@@ -1,5 +1,6 @@
 <script setup>
 import AppLayout from '@/Layouts/AppLayout.vue';
+import { Link } from '@inertiajs/vue3';
 import { computed, onMounted, ref } from 'vue';
 
 const props = defineProps({
@@ -7,7 +8,7 @@ const props = defineProps({
 });
 
 const headers = ref([
-    { title: 'Source', key: 'name' },
+    { title: 'Source', key: 'id' },
     { title: 'Last import', key: 'lastImport.started_at' },
     { title: 'Last import status', key: 'lastImport.status' },
     { key: 'actions' },
@@ -17,7 +18,7 @@ const ajaxing = ref(false);
 
 const sinksKeyed = computed(() => {
     const ret = {};
-    props.sinks.forEach((sink) => ret[sink.name] = sink);
+    props.sinks.forEach((sink) => ret[sink.id] = sink);
     return ret;
 });
 
@@ -25,13 +26,13 @@ function naIfEmpty(item, key) {
     return item.columns[key] || 'N/A';
 }
 
-function importSingle(sinkName) {
+function importSingle(sinkId) {
     ajaxing.value = true;
     axios.post(
-        'api/sink/import',
-        { sink_name: sinkName }
+        'api/sink',
+        { sink_id: sinkId }
     ).then((result) => {
-        sinksKeyed.value[sinkName].lastImport = result.data;
+        sinksKeyed.value[sinkId].lastImport = result.data;
     }).catch((error) => {
         console.warn(error.response.data);
     }).finally(() => ajaxing.value = false);
@@ -40,7 +41,7 @@ function importSingle(sinkName) {
 onMounted(() => {
     Echo.private('App.Models.SinkImport').listen('.SinkImportUpdated', (event) => {
         const sinkImport = event.model;
-        sinksKeyed.value[sinkImport.sink_name].lastImport = sinkImport;
+        sinksKeyed.value[sinkImport.sink_id].lastImport = sinkImport;
     });
 });
 
@@ -52,11 +53,16 @@ onMounted(() => {
       :headers="headers"
       :items="sinks"
       items-per-page="100"
-      item-value="name"
+      item-value="id"
       no-filter
     >
       <template #item.lastImport.status="{ item }">
         {{ naIfEmpty(item, 'lastImport.status') }}
+      </template>
+      <template #item.id="{ item }">
+        <Link :href="`/sink/${item.value}`">
+          {{ item.raw.title }}
+        </Link>
       </template>
       <template #item.actions="{ item }">
         <v-btn icon flat @click="importSingle(item.value)">