+++ resources/js/plugins/vuetify.js
@@ -8,12 +8,12 @@
 import '@mdi/font/css/materialdesignicons.css';
 import 'vuetify/styles';
 import { createVuetify } from 'vuetify';
-import { VDataTable } from 'vuetify/labs/VDataTable';
+import { VDataTable, VDataTableServer } from 'vuetify/labs/VDataTable';
 import { aliases, mdi } from 'vuetify/iconsets/mdi';
 
 // https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
 export default createVuetify({
-    components: { VDataTable },
+    components: { VDataTable, VDataTableServer },
     icons: {
         defaultSet: 'mdi',
         aliases,