+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\ChunkController;
 use App\Http\Controllers\ImportController;
 use App\Http\Controllers\UserRoleController;
 use Illuminate\Http\Request;
@@ -20,10 +21,16 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::resource('sink/import', ImportController::class)->only(['store', 'show', 'update', 'destroy']);
-
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('usersWithRoles', [UserRoleController::class, 'getUsersWithRoles']);
     Route::get('userRolesWithPermissions', [UserRoleController::class, 'getRolesAndPermissions']);
     Route::post('updateUserRole/{id}/{role}', [UserRoleController::class, 'setUserRole']);
+});
+
+Route::resource('sink', ImportController::class)->only(['store', 'show', 'update', 'destroy']);
+Route::resource('sink/{sinkId}/chunk', ChunkController::class)->except(['destroy', 'store']);
+
+Route::controller(ChunkController::class)->group(function () {
+    Route::get('sink/{sinkId}/chunk', 'index');
+    Route::post('sink/{sinkId}/chunk/fetch', 'fetch');
 });